♻️ refactor: use external flag instead of internal flag
